+++ radix/tree.go
@@ -16,7 +16,7 @@ func NewTree() (result *Tree) {
 	result.root.rehash(nil)
 	return
 }
-func newTreeWith(key []nibble, value Hasher, version uint32) (result *Tree) {
+func newTreeWith(key []nibble, value Hasher, version int64) (result *Tree) {
 	result = NewTree()
 	result.PutVersion(key, value, 0, version)
 	return
@@ -83,7 +83,7 @@ func (self *Tree) del(key []nibble) (old Hasher, existed bool) {
 	return
 }
 
-func (self *Tree) getSubTree(key []nibble) (subTree *Tree, version uint32) {
+func (self *Tree) getSubTree(key []nibble) (subTree *Tree, version int64) {
 	var value Hasher
 	var existed bool
 	if value, version, existed = self.root.get(key); existed {
@@ -125,11 +125,11 @@ func (self *Tree) SubDel(key, subKey []byte) (old Hasher, existed bool) {
 func (self *Tree) Finger(key []nibble) (result *Print) {
 	return self.root.finger(&Print{nil, nil, 0, false, nil}, key)
 }
-func (self *Tree) GetVersion(key []nibble) (value Hasher, version uint32, existed bool) {
+func (self *Tree) GetVersion(key []nibble) (value Hasher, version int64, existed bool) {
 	value, version, existed = self.root.get(key)
 	return
 }
-func (self *Tree) PutVersion(key []nibble, value Hasher, expected, version uint32) {
+func (self *Tree) PutVersion(key []nibble, value Hasher, expected, version int64) {
 	if _, current, existed := self.root.get(key); !existed || current == expected {
 		self.root, _, _, existed = self.root.insert(nil, false, newNode(key, value, version, true))
 		if !existed {
@@ -137,7 +137,7 @@ func (self *Tree) PutVersion(key []nibble, value Hasher, expected, version uint3
 		}
 	}
 }
-func (self *Tree) DelVersion(key []nibble, expected uint32) {
+func (self *Tree) DelVersion(key []nibble, expected int64) {
 	if _, current, existed := self.root.get(key); existed && current == expected {
 		var existed bool
 		self.root, _, existed = self.root.del(nil, key)
@@ -147,19 +147,19 @@ func (self *Tree) DelVersion(key []nibble, expected uint32) {
 	}
 }
 
-func (self *Tree) SubFinger(key, subKey []nibble, expected uint32) (result *Print) {
+func (self *Tree) SubFinger(key, subKey []nibble, expected int64) (result *Print) {
 	if subTree, subTreeVersion := self.getSubTree(key); subTree != nil && subTreeVersion == expected {
 		result = subTree.Finger(subKey)
 	}
 	return
 }
-func (self *Tree) SubGetVersion(key, subKey []nibble, expected uint32) (value Hasher, version uint32, existed bool) {
+func (self *Tree) SubGetVersion(key, subKey []nibble, expected int64) (value Hasher, version int64, existed bool) {
 	if subTree, subTreeVersion := self.getSubTree(key); subTree != nil && subTreeVersion == expected {
 		value, version, existed = subTree.GetVersion(subKey)
 	}
 	return
 }
-func (self *Tree) SubPutVersion(key, subKey []nibble, value Hasher, expected, subExpected, subVersion uint32) {
+func (self *Tree) SubPutVersion(key, subKey []nibble, value Hasher, expected, subExpected, subVersion int64) {
 	if subTree, subTreeVersion := self.getSubTree(key); subTree == nil || subTreeVersion == expected {
 		if subTree == nil {
 			subTree = newTreeWith(subKey, value, subVersion)
@@ -170,7 +170,7 @@ func (self *Tree) SubPutVersion(key, subKey []nibble, value Hasher, expected, su
 	}
 	return
 }
-func (self *Tree) SubDelVersion(key, subKey []nibble, expected, subExpected uint32) {
+func (self *Tree) SubDelVersion(key, subKey []nibble, expected, subExpected int64) {
 	if subTree, subTreeVersion := self.getSubTree(key); subTree != nil && subTreeVersion == expected {
 		subTree.DelVersion(subKey, subExpected)
 		if subTree.Size() == 0 {
